refactor: move new components to `src/components/` folder
